+++ devito/finite_differences/__init__.py
@@ -1,3 +1,4 @@
+from .lazy import *  # noqa
 from .differentiable import *  # noqa
 from .finite_difference import *  # noqa
 from .derivative import *  # noqa
