+++ devito/tools/abc.py
@@ -2,7 +2,7 @@ import abc
 from hashlib import sha1
 
 
-__all__ = ['Tag', 'Signer', 'Pickable', 'Evaluable', 'Singleton']
+__all__ = ['Tag', 'Signer', 'Pickable', 'Singleton']
 
 
 class Tag(abc.ABC):
@@ -148,47 +148,6 @@ class Pickable(object):
     def __getnewargs_ex__(self):
         return (tuple(getattr(self, i) for i in self._pickle_args),
                 {i.lstrip('_'): getattr(self, i) for i in self._pickle_kwargs})
-
-
-class Evaluable(object):
-
-    """
-    A mixin class for types that may carry nested unevaluated arguments.
-
-    This mixin is useful to implement systems based upon lazy evaluation.
-    """
-
-    @classmethod
-    def _evaluate_maybe_nested(cls, maybe_evaluable):
-        if isinstance(maybe_evaluable, Evaluable):
-            return maybe_evaluable.evaluate
-        try:
-            # Not an Evaluable, but some Evaluables may still be hidden within `args`
-            if maybe_evaluable.args:
-                evaluated = [Evaluable._evaluate_maybe_nested(i)
-                             for i in maybe_evaluable.args]
-                return maybe_evaluable.func(*evaluated)
-            else:
-                return maybe_evaluable
-        except AttributeError:
-            # No `args` to be visited
-            return maybe_evaluable
-
-    @property
-    def args(self):
-        return ()
-
-    @property
-    def func(self):
-        return self.__class__
-
-    def _evaluate_args(self):
-        return [Evaluable._evaluate_maybe_nested(i) for i in self.args]
-
-    @property
-    def evaluate(self):
-        """Return a new object from the evaluation of ``self``."""
-        return self.func(*self._evaluate_args())
 
 
 class Singleton(type):
