+++ devito/types/equation.py
@@ -4,8 +4,8 @@ import sympy
 
 from cached_property import cached_property
 
-from devito.finite_differences import default_rules
-from devito.tools import Evaluable, as_tuple
+from devito.finite_differences import Evaluable, default_rules
+from devito.tools import as_tuple
 
 __all__ = ['Eq', 'Inc', 'solve']
